restructures examples and some functions
